+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 **Hearthstone**
 - Fixed The Kingslayer not Highlighting Incindius' Eruptions.
 - Added cards offered by Waveshaping to the bottom of the deck.
+- Added cards discovered by Kaldorei Cultivator to the bottom of the deck.
 
 
 ## **Release v1.48.2 - 2025-10-31**